farm-hustle-backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.31
psycopg2-binary==2.9.9
pydantic==2.7.4
pydantic-settings==2.3.3
python-multipart==0.0.9
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
